Add generated typescript definitions as part of build
Adds a new script `build-types`, which generates typescript definitions for the client class (using a similar pattern to `build-rest`).

Note: this is not comprehensive. Response types are only partially defined (they aren't in the existing json schema), and the plugins are not defined, but this still greatly improves the explorability and autocomplete of the API in a TS environment.